move read loop and strategy to their own files
