+++ src/orbitalcoms/coms/drivers/basedriver.py
@@ -1,25 +1,17 @@
 from __future__ import annotations
 
-import multiprocessing as mp
-from multiprocessing.connection import Connection
-from threading import Condition, Event, Thread
-from typing import TYPE_CHECKING, Protocol, Set, Tuple
+from threading import Condition
+from typing import TYPE_CHECKING, Set
 
 from ..errors import ComsDriverReadError, ComsDriverWriteError
 from ..messages import construct_message
 from ..subscribers import OneTimeComsSubscription
+from .driverreadloop import ComsDriverReadLooop
 
 if TYPE_CHECKING:
     from ..messages import ComsMessage, ParsableComType
+    from ..strategies import ComsStrategy
     from ..subscribers import ComsSubscriberLike
-
-
-class ComsStrategy(Protocol):
-    def read(self) -> ComsMessage:
-        ...
-
-    def write(self, m: ComsMessage) -> None:
-        ...
 
 
 class BaseComsDriver:
@@ -97,51 +89,3 @@ class BaseComsDriver:
                 # TODO: Add logging here
                 if not s.expect_err:
                     self.unregister_subscriber(s)
-
-
-class ComsDriverReadLooop(Thread):
-    def __init__(
-        self,
-        coms: BaseComsDriver,
-        name: str | None = None,
-        daemon: bool | None = None,
-    ) -> None:
-        super().__init__(name=name, daemon=daemon)
-        self._stop_event = Event()
-        self._coms = coms
-
-    def run(self) -> None:
-        proc, conn = self._spawn_get_msg_proc()
-        proc.start()
-
-        while not self._stop_event.is_set():
-            if not proc.is_alive():
-                recived = conn.recv()
-                if isinstance(recived, Exception):
-                    # TODO: Add logging
-                    ...
-                else:
-                    self._coms._notify_subscribers(recived)
-                proc, conn = self._spawn_get_msg_proc()
-                proc.start()
-            proc.join(timeout=1)
-
-        if proc.is_alive():
-            proc.terminate()
-
-    def _spawn_get_msg_proc(
-        self,
-    ) -> Tuple[mp.Process, Connection]:
-        a, b = mp.Pipe()
-
-        def get_msg(conn: Connection) -> None:
-            try:
-                conn.send(self._coms.strategy.read())
-            except Exception as e:
-                conn.send(e)
-
-        return mp.Process(target=get_msg, args=(a,), daemon=True), b
-
-    def stop(self, timeout: float | None = None) -> None:
-        self._stop_event.set()
-        self.join(timeout=timeout)
