+++ src/orbitalcoms/coms/strategies/__init__.py
@@ -1,4 +1,5 @@
 from .serialcomsstrategy import SerialComsStrategy
 from .socketcomsstrategy import SocketComsStrategy
+from .strategy import ComsStrategy
 
-__all__ = ["SerialComsStrategy", "SocketComsStrategy"]
+__all__ = ["SerialComsStrategy", "SocketComsStrategy", "ComsStrategy"]
