+++ src/orbitalcoms/coms/strategies/strategy.py
@@ -1,0 +1,11 @@
+from typing import Protocol
+
+from ..messages.message import ComsMessage
+
+
+class ComsStrategy(Protocol):
+    def read(self) -> ComsMessage:
+        ...
+
+    def write(self, m: ComsMessage) -> None:
+        ...
